Fake Travis alert?

+++ htdocs/eventorganization/conferenceorboothattendee_list.php
@@ -278,7 +278,7 @@ if ($massaction == 'confirm_premassmail' && $permissiontoread) {
 					$verified_mailings[] = $mailingid;
 					dol_syslog('Using mailing->fetch() to verify mailingid='.$mailingid, LOG_DEBUG);
 				} else {
-					dol_syslog('No '.$checkProjectAccessRight.' access OR user='.$user->id.' is not a Contact to projectid='.$mailing_fk_project.' in mailingid='.$mailingid.' in array select_mailing on page eventorganization/conferenceorboothattendee_list.php massaction confirm_premassmail', LOG_ERR);
+					dol_syslog('No '.$checkProjectAccessRight.' access OR user='.((int) $user->id).' is not a Contact to projectid='.$mailing_fk_project.' in mailingid='.$mailingid.' in array select_mailing on page eventorganization/conferenceorboothattendee_list.php massaction confirm_premassmail', LOG_ERR);
 					$verified_mailings = null;
 					$button_add_mailing = null;
 					$button_delete_mailing = null;
@@ -439,7 +439,7 @@ $parameters = array();
 $reshook = $hookmanager->executeHooks('printFieldListFrom', $parameters, $object, $action); // Note that $action and $object may have been modified by hook
 $sql .= $hookmanager->resPrint;
 if ($object->ismultientitymanaged == 1) {
-	$sql .= " WHERE t.entity IN (".getEntity($object->element).")";
+	$sql .= " WHERE t.entity IN (".((int) getEntity($object->element)).")";
 } else {
 	$sql .= " WHERE 1 = 1";
 }
